Added code to dynamically handle new regions
