+++ plugins/elite.js
@@ -11,6 +11,7 @@ var alphanum = require("../server/alphanum.js");
 var regionjpg = require("./elite/regionjpg.js");
 var gmp = require('./elite/gmp.js');
 var regions = require('./elite/regions.js');
+var elitelib = require('./elite/elitelib.js');
 
 var botcfg = null;
 var pmIfSpam = false;
@@ -21,79 +22,6 @@ try{
     cmdrs = require(path.resolve(base.path, "cmdrs.json"));
 } catch(e) {
     cmdrs = {};
-}
-
-var NEW_THRESHHOLD = (7 * 24 * 60 * 60 * 1000);
-
-var densitySigmaArray = [ // in SI units of kg/m^3
-        {likelyType: "IW", densities: [1.06e+3, 1.84e+3, 2.62e+3, 3.40e+3]},
-        {likelyType: "RIW", densities: [2.25e+3, 2.82e+3, 3.38e+3, 3.95e+3]},
-        {likelyType: "RW", densities: [2.94e+3, 3.77e+3, 4.60e+3, 5.43e+3]},
-        {likelyType: "HMC", densities: [1.21e+3, 4.60e+3, 8.00e+3, 1.14e+4]},
-        {likelyType: "MR", densities: [1.47e+3, 7.99e+3, 1.45e+4, 2.10e+4]},
-        {likelyType: "WW", densities: [1.51e+3, 4.24e+3, 6.97e+3, 9.70e+3]},
-        {likelyType: "ELW", densities: [4.87e+3, 5.65e+3, 6.43e+3, 7.21e+3]},
-        {likelyType: "AW", densities: [4.23e+2, 3.50e+3, 6.59e+3, 9.67e+3]}
-];
-
-function handleGravity(planetMass, planetRadius) {
-    var G = 6.67e-11;
-    var earthMass = 5.98e24;
-    var earthRadius = 6367444.7;
-    var baseG = G * earthMass / (earthRadius * earthRadius);
-    var planetG = G * planetMass * earthMass / (planetRadius * planetRadius * 1e6);
-    var planetDensity = planetMass * earthMass / (4.0 / 3.0 * Math.PI * planetRadius * planetRadius * planetRadius) * 1e-9; // in SI units of kg/m^3
-    var planetM2Str = planetG.toFixed(2);
-    var planetGStr = (planetG / baseG).toFixed(2);
-    var planetDensityStr = planetDensity.toExponential(2);
-    var maybeTypes = [];
-    var likelyTypes = [];
-
-    for (var i = 0; i < densitySigmaArray.length; i++) {
-        var row = densitySigmaArray[i];
-        if (planetDensity > row.densities[1] && planetDensity < row.densities[2]) {
-            likelyTypes.push(row.likelyType);
-        } else if (planetDensity > row.densities[0] && planetDensity < row.densities[3]) {
-            maybeTypes.push(row.likelyType);
-        }
-    }
-    var densityString = "";
-    if (likelyTypes.length > 0) {
-        densityString += "\n**Likely**: " + likelyTypes.sort().join(", ");
-    }
-    if (maybeTypes.length > 0) {
-        densityString += "\n**Possible**: " + maybeTypes.sort().join(", ");
-    }
-
-    var ret = "The gravity for a planet with " + planetMass + " Earth masses and a radius of ";
-    ret += planetRadius + " km is **" + planetM2Str + "** m/s^2 or **" + planetGStr;
-    ret += "** g. It has a density of **" + planetDensityStr + "** kg/m^3." + densityString;
-    return ret;
-}
-
-/**
-Implements Jackie Silver's method of stellar density calculation
-
-To find the density, look at the navpanel. If there are more than 50 star systems within 20 light years, 
-only the first 50 systems will be shown. Otherwise, the navpanel will show all systems within 20 light years.
-
-This gives us two ways of finding the density:
-
-1) For dense areas where there are more than 50 star systems, look at the last star system in the list, 
-and see what its distance ("r") from your ship is. This lets us estimate the density as rho = 50 / ((4pi/3) * (r^3))
-
-2) For sparse areas where there are less than 50 star systems, count how many star systems ("n") are visible in 
-total in the navpanel. This lets us estimate the density as rho = n / ((4pi/3) * (20^3))
-
-This function prioritizes 1) over 2) - so it will perform method 1 if r is not undefined, otherwise 2
-**/
-function calcRho(r, n) {
-    if (r) {
-        return 50 / ((4 * Math.PI / 3) * Math.pow(r, 3));
-    } else if (n) {
-        return n / ((4 * Math.PI / 3) * Math.pow(20, 3));
-    }
-    return undefined;
 }
 
 function writeAliases() {
@@ -109,8 +37,7 @@ function writeCmdrs() {
 }
 
 function getRegionMap(location, callback) {
-    var names = location.split(/ [a-z][a-z]-[a-z] /i);
-    var key = names[0].toLowerCase();
+    var key = elitelib.getRegionName(location);
     regions.getRegionByKey(key, function(region) {
         if (region) {
             var name = region.region;
@@ -123,7 +50,7 @@ function getRegionMap(location, callback) {
                 if (sys && sys.coords) {
                     var coords = sys.coords;
                     // z is the Y coordinate looking down at the map
-                    regionjpg.fetchRegionMapByCoords(coords.x, coords.z, function(rgn) {
+                    regionjpg.fetchRegionMapByCoords(coords.x, coords.z, location, function(rgn) {
                         callback(rgn);
                     });
                 } else {
@@ -146,12 +73,6 @@ function _showRegion(region, bot, msg) {
                 regionString = orgRegion + " (" + region + ")";
             }
             if (data.map) {
-                // regionString = data.region;
-                var newRegionDate = new Date().getTime() - NEW_THRESHHOLD;
-                var regionDate = new Date(data.date).getTime();
-                if (regionDate > newRegionDate) {
-                    regionString += "\n*Newly trilaterated region: " + data.date + "*";
-                }
                 msg.channel.sendFile(path.resolve(base.path, "plugins/elite/maps/" + data.map), data.map, regionString);
             } else {
                 utils.sendMessage(bot, msg.channel, "Sorry, I have no map for " + regionString);
@@ -356,7 +277,7 @@ var commands = {
                 var planetRadius = +(args[2]);
                 if(!isNaN(planetMass) && !isNaN(planetRadius)) {
                     displayUsage = false;
-                    utils.sendMessage(bot, msg.channel, handleGravity(planetMass, planetRadius));
+                    utils.sendMessage(bot, msg.channel, elitelib.handleGravity(planetMass, planetRadius));
                 }
             }
             if (displayUsage) {
@@ -372,7 +293,7 @@ var commands = {
             var displayUsage = true;
             if (args.length === 2) {
                 var distance = args[1];
-                var density = calcRho(distance, undefined);
+                var density = elitelib.calcRho(distance, undefined);
                 if(!isNaN(density)) {
                     displayUsage = false;
                     utils.sendMessage(bot, msg.channel, "rho = " + density);
@@ -391,7 +312,7 @@ var commands = {
             var displayUsage = true;
             if (args.length === 2) {
                 var count = args[1];
-                var density = calcRho(undefined, count);
+                var density = elitelib.calcRho(undefined, count);
                 if(!isNaN(density)) {
                     displayUsage = false;
                     utils.sendMessage(bot, msg.channel, "rho = " + density);
