+++ plugins/elite/regionjpg.js
@@ -4,6 +4,7 @@ var fs = require("fs");
 var gm = require('gm');
 var imageMagick = gm.subClass({ imageMagick: true });
 var regions = require('./regions.js');
+var elitelib = require('./elitelib.js');
 
 // Magic numbers based on the source image, Galaxy.jpg
 const KLY_TO_PIXEL = 1000 / 12.5;
@@ -81,16 +82,21 @@ var generateRegionMapByCoords = function(x, y, name, filename, callback) {
     });
 }
 
-var fetchRegionMapByCoords = function(x, y, callback) {
-    console.log("fetchRegionMapByCoords(" + x + ", " + y + ", callback)");
+var fetchRegionMapByCoords = function(x, y, location, callback) {
+    console.log("fetchRegionMapByCoords(" + x + ", " + y + ", " + location + ", callback)");
     x = normalizeCoordX(x);
     y = normalizeCoordY(y);
     var filename = generateCoordFileName(x, y);
+    var name = undefined;    // This is only set for regions
+    if (elitelib.isProcGen(location)) {
+        name = elitelib.getRegionName(location);
+        filename = name.toLowerCase();
+    }
     regions.getRegionByKey(filename, function(rgn) {
         if (rgn && rgn.map && (fs.existsSync(_destDir + rgn.map))) {
             callback(rgn);
         } else {
-            generateRegionMapByCoords(x, y, undefined, filename, callback);
+            generateRegionMapByCoords(x, y, name, filename, callback);
         }
     });
 }
